requirements.txt
asgiref==3.8.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.1.7
django-allauth==65.13.0
django-environ==0.12.0
django-jazzmin==3.0.1
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.11
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
rest-framework-simplejwt==0.0.2
setuptools==75.8.0
sqlparse==0.5.3
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.11.0
